Rename Context folder to Persistent layer.
